+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 1.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Compatibility fixes with aiohttp 2
+  [vangheem]
 
 
 1.0.0 (2017-04-24)